[gpt_reco_app] organize legal pages
