+++ gpt_reco_app/src/pages/legal/PrivacyPolicy.jsx
