+++ gpt_reco_app/src/pages/legal/TermsOfService.jsx
